requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
feedparser>=6.0
playwright>=1.49
Pillow>=11.0
anthropic>=0.42
python-dotenv>=1.0
schedule>=1.2
click>=8.1
